Adding router/controller for staff page
