+++ app.js
@@ -3,12 +3,15 @@ const express = require('express');     //  Used for simplified code expressions
 const mongoose = require('mongoose');   //  Used to access the cloud db
 const morgan = require('morgan');       //  Used to log requests and responses in the terminal when running in development environment
 const { render } = require('ejs');      //  Used to embed javascript in the html
-const cookieParser = require('cookie-parser')   //  Used to read cookies
-const authRoutes = require('./routes/authRoutes');     
-const { requireAuth, checkUser } = require('./middleware/authMiddleware');
+const cookieParser = require('cookie-parser')   //  Used to read cookies    
+
+// Declare authorization middleware
+const { requireAuth, checkUser, checkTeacher } = require('./middleware/authMiddleware');
 
 // Import Routers
+const authRoutes = require('./routes/authRoutes'); 
 const courseRoutes = require('./routes/courseRoutes');
+const staffRoutes = require('./routes/staffRoutes');
 
 // DB Models
 const Course = require('./models/Course');
@@ -71,26 +74,7 @@ app.get('/students', requireAuth, (req, res) => {
 });
 
 // Page for staff to add and modify courses
-app.get('/staff', (req, res) => {
-    let subjects = {};
-
-    Subject.find().sort({ title: 1 })
-        .then(sResult => {
-            subjects = sResult;            
-        })
-        .then(() => {
-            Course.find().sort({ cname: 1 })
-                .then((cResult) => {                    
-                    res.render('staff', {courses: cResult, subjects: subjects, title: 'Staff' });
-                })
-                .catch((err) => {
-                    console.log(err);
-                });
-            })                
-        .catch((err) => {
-            console.log(err);
-        });
-});
+app.use('/staff', staffRoutes);
 
 app.use('/courses', courseRoutes);
 app.use(authRoutes);
